requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==4.0.3
django-cleanup==6.0.0
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
idna==3.3
Pillow==9.1.0
psycopg2==2.9.3
PyJWT==2.3.0
python-dateutil==1.5
pytz==2022.1
requests==2.27.1
sentry-sdk==1.5.8
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0
